+++ packages/jco-std/package.json
@@ -27,20 +27,20 @@
   ],
   "type": "module",
   "exports": {
-    "./node24.x/assert": {
-      "types": "./dist/node24.x/assert/index.d.ts",
-      "browser": "./dist/node24.x/assert/index.js",
-      "default": "./dist/node24.x/assert/index.js"
+    "./wasi/0.2.x/node/24.x.x/assert": {
+      "types": "./dist/wasi/0.2.x/node/24.x.x/assert/index.d.ts",
+      "browser": "./dist/wasi/0.2.x/node/24.x.x/assert/index.js",
+      "default": "./dist/wasi/0.2.x/node/24.x.x/assert/index.js"
     },
-    "./node24.x/path": {
-      "types": "./dist/node24.x/path.d.ts",
-      "browser": "./dist/node24.x/path.js",
-      "default": "./dist/node24.x/path.js"
+    "./wasi/0.2.x/node/24.x.x/path": {
+      "types": "./dist/wasi/0.2.x/node/24.x.x/path.d.ts",
+      "browser": "./dist/wasi/0.2.x/node/24.x.x/path.js",
+      "default": "./dist/wasi/0.2.x/node/24.x.x/path.js"
     },
     "./node/path": {
-      "types": "./dist/node24.x/path.d.ts",
-      "browser": "./dist/node24.x/path.js",
-      "default": "./dist/node24.x/path.js"
+      "types": "./dist/wasi/0.2.x/node/24.x.x/path.d.ts",
+      "browser": "./dist/wasi/0.2.x/node/24.x.x/path.js",
+      "default": "./dist/wasi/0.2.x/node/24.x.x/path.js"
     },
     "./wasi/0.2.x/http/adapters/hono/server": {
       "types": "./dist/wasi/0.2.12/http/adapters/hono/server.d.ts",
