+++ packages/jco-std/test/wasi/0.2.x/node/24.x.x/helpers/assert.ts
@@ -1,6 +1,6 @@
 import nodeAssert from "node:assert";
 import { describe, expect } from "vitest";
-import { AssertionError } from "../../../src/node24.x/assert/index.js";
+import { AssertionError } from "../../../../../../src/wasi/0.2.x/node/24.x.x/assert/index.js";
 
 export interface CapturedResult {
   passed: boolean;
