+++ packages/jco/test/node/assert.js
@@ -7,7 +7,7 @@ import { COMPONENT_JS_FIXTURES_DIR } from "../common.js";
 import { exec, getTmpDir, jcoPath } from "../helpers.js";
 
 suite("node:assert", () => {
-    // TODO(unskip): jco pins @bytecodealliance/jco-std/node24.x/assert, and no published
+    // TODO(unskip): jco pins @bytecodealliance/jco-std/wasi/0.2.x/node/24.x.x/assert, and no published
     // jco-std exports the assert shim at all -- it only exists in the workspace copy. Unskip once
     // a jco-std release carrying it is published and jco's range is bumped to it.
     test.skip("bundles and executes APIs guest-side", async () => {
