+++ packages/jco/test/node/path.js
@@ -10,9 +10,9 @@ import { assert, expect, suite, test } from "vitest";
 import { componentizeFixture, setupAsyncTest } from "../helpers.js";
 
 suite("node:path in a component", () => {
-    // TODO(unskip): jco pins @bytecodealliance/jco-std/node24.x/path, which the published
+    // TODO(unskip): jco pins @bytecodealliance/jco-std/wasi/0.2.x/node/24.x.x/path, which the published
     // jco-std does not export yet -- it only has the ./node/path alias. Unskip once a jco-std
-    // release carrying the node24.x entry points is published and jco's range is bumped to it.
+    // release carrying the wasi/0.2.x/node/24.x.x entry points is published and jco's range is bumped to it.
     test.skip("componentizes and runs lexical and cwd-backed path operations", async () => {
         const { componentPath, stderr } = await componentizeFixture({ fixture: "node-path", bundle: true });
         assert.strictEqual(stderr, "");
